requirements.txt
Django>=3.0
psycopg2==2.8.4
social-auth-app-django==3.1.0
djangorestframework==3.11.0
django-iban-field==0.8